Moved depthTraversable functions into their own property
